Update HiveClient.py
Hi Miklos, thank you for the constructive feedback. I cleansed the code and updated commit. 

Problem: 

The table exports that does not have location or path in ddl definition fails at destination fails during import. I added a fix to take care of this. 

Here is the workspace with mixed set of ddls with variety of tables, managed, external, csv, orc, delta, parquet and hive table. 

https://adb-5985677774820244.4.azuredatabricks.net/?o=5985677774820244#setting/clusters

If the tables are exported in this workspace with current migrate script 

1) it would fail at destination with empty state. I add Charles Bernard as witness. charles.bernard@databricks.com and we had to make this change to migrate Prudencial at APAC. 
2) With the patch this is fixed.